refactor: 내가ë 북마크한 코스 목록 조회 리팩터
