+++ src/main/java/com/example/wherewego/domain/common/enums/CourseTheme.java
@@ -1,4 +1,4 @@
-package com.example.wherewego.common.enums;
+package com.example.wherewego.domain.common.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
