+++ src/main/java/com/example/wherewego/domain/courses/entity/Rating.java
@@ -1,8 +1,17 @@
 package com.example.wherewego.domain.courses.entity;
 
-import com.example.wherewego.common.entity.BaseEntity;
+import com.example.wherewego.domain.common.entity.BaseEntity;
 import com.example.wherewego.domain.user.entity.User;
-import jakarta.persistence.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -12,29 +21,29 @@ import lombok.NoArgsConstructor;
 @Table(name = "ratings")
 public class Rating extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "course_id", nullable = false)
-    private Course course;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "course_id", nullable = false)
+	private Course course;
 
-    @Column(nullable = false)
-    private int rating;
+	@Column(nullable = false)
+	private int rating;
 
-    public Rating(User user, Course course, int rating) {
-        this.user = user;
-        this.course = course;
-        this.rating = rating;
-    }
+	public Rating(User user, Course course, int rating) {
+		this.user = user;
+		this.course = course;
+		this.rating = rating;
+	}
 
-    //수정용
-    public void updateRating(int rating) {
-        this.rating = rating;
-    }
+	//수정용
+	public void updateRating(int rating) {
+		this.rating = rating;
+	}
 }
