+++ src/main/java/com/example/wherewego/domain/courses/service/CourseLikeService.java
@@ -3,7 +3,7 @@ package com.example.wherewego.domain.courses.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.wherewego.common.enums.ErrorCode;
+import com.example.wherewego.domain.common.enums.ErrorCode;
 import com.example.wherewego.domain.courses.dto.response.CourseLikeResponseDto;
 import com.example.wherewego.domain.courses.entity.Course;
 import com.example.wherewego.domain.courses.entity.Like;
@@ -27,7 +27,7 @@ public class CourseLikeService {
 		User user = userService.getUserById(userId);
 		Course course = courseService.getCourseById(courseId);
 		// 좋아요 중복 검사
-		if(likeRepository.existsByUserIdAndCourseId(userId, courseId)) {
+		if (likeRepository.existsByUserIdAndCourseId(userId, courseId)) {
 			throw new CustomException(ErrorCode.LIKE_ALREADY_EXISTS);
 		}
 		// 레파지토리 저장
@@ -37,9 +37,9 @@ public class CourseLikeService {
 		course.incrementLikeCount();
 		// 반환
 		return new CourseLikeResponseDto(
-				savedLike.getId(),
-				savedLike.getUser().getId(),
-				savedLike.getCourse().getId()
+			savedLike.getId(),
+			savedLike.getUser().getId(),
+			savedLike.getCourse().getId()
 		);
 	}
 
@@ -47,7 +47,7 @@ public class CourseLikeService {
 	public void courseLikeDelete(Long userId, Long courseId) {
 		// 좋아요 존재 검사
 		Like like = likeRepository.findByUserIdAndCourseId(userId, courseId)
-				.orElseThrow(() -> new CustomException(ErrorCode.LIKE_NOT_FOUND));
+			.orElseThrow(() -> new CustomException(ErrorCode.LIKE_NOT_FOUND));
 		// hard delete
 		likeRepository.delete(like);
 		// 코스 테이블의 좋아요 수(like_count) -1
