+++ src/main/java/com/example/wherewego/global/exception/CustomException.java
@@ -1,15 +1,16 @@
 package com.example.wherewego.global.exception;
 
-import com.example.wherewego.common.enums.ErrorCode;
+import com.example.wherewego.domain.common.enums.ErrorCode;
+
 import lombok.Getter;
 
 @Getter
 public class CustomException extends RuntimeException {
-  private final ErrorCode errorCode;
+	private final ErrorCode errorCode;
 
-  public CustomException(ErrorCode errorCode) {
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
+	public CustomException(ErrorCode errorCode) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+	}
 
 }
