+++ src/test/java/com/example/wherewego/domain/course/service/CourseBookmarkServiceTest.java
@@ -13,8 +13,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 
-import com.example.wherewego.common.enums.ErrorCode;
 import com.example.wherewego.domain.auth.Provider;
+import com.example.wherewego.domain.common.enums.ErrorCode;
 import com.example.wherewego.domain.courses.dto.response.CourseBookmarkResponseDto;
 import com.example.wherewego.domain.courses.entity.Course;
 import com.example.wherewego.domain.courses.entity.CourseBookmark;
