+++ src/test/java/com/example/wherewego/domain/course/service/CourseRatingServiceTest.java
@@ -12,8 +12,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.example.wherewego.common.enums.ErrorCode;
 import com.example.wherewego.domain.auth.Provider;
+import com.example.wherewego.domain.common.enums.ErrorCode;
 import com.example.wherewego.domain.courses.dto.request.CourseRatingRequestDto;
 import com.example.wherewego.domain.courses.dto.response.CourseRatingResponseDto;
 import com.example.wherewego.domain.courses.entity.Course;
